♻️ Move DraggableBoundingBox alongside other cropper-specific files
