+++ src/routes/(app)/(sidepanel)/o/[observation]/crop/[image]/DraggableBoundingBox.test.svelte.js
